# examples/taskflow-api/logos/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  version: 0.1.0
  description: |
    Reference demo API for OpenLogos — JWT auth and per-user tasks (SQLite).

servers:
  - url: http://localhost:3000
    description: Local development

tags:
  - name: auth
  - name: tasks

paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
        '409':
          description: Email already registered

  /auth/login:
    post:
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
        '401':
          description: Invalid credentials

  /tasks:
    post:
      tags: [tasks]
      summary: Create a task
      security: [bearerAuth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Validation error
        '401':
          description: Unauthorized

    get:
      tags: [tasks]
      summary: List tasks for the current user
      security: [bearerAuth]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [tasks]
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized

  /tasks/{id}:
    get:
      tags: [tasks]
      summary: Get a task by id
      security: [bearerAuth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid id
        '401':
          description: Unauthorized
        '404':
          description: Not found

    patch:
      tags: [tasks]
      summary: Update a task
      security: [bearerAuth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchTaskRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Not found

    delete:
      tags: [tasks]
      summary: Delete a task
      security: [bearerAuth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No content
        '400':
          description: Invalid id
        '401':
          description: Unauthorized
        '404':
          description: Not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          type: object
          required: [id, email]
          properties:
            id:
              type: integer
            email:
              type: string
        token:
          type: string

    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string

    PatchTaskRequest:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [pending, done]

    Task:
      type: object
      required: [id, userId, title, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [pending, done]
        createdAt:
          type: string
          format: date-time